// packages/ms-teams-webhook/src/adaptive-card.d.ts
export type Action = Extendable & {
  /**
   * Must be `Action.Execute`
   */
  type?: "Action.Execute";
  /**
   * The card author-defined verb associated with this action.
   */
  verb?: string;
  /**
   * Initial data that input fields will be combined with. These are essentially ‘hidden’ properties.
   */
  data?:
    | string
    | {
        [k: string]: unknown;
      };
  /**
   * Controls which inputs are associated with the action.
   */
  associatedInputs?:
    | ("Auto" | "None")
    | {
        [k: string]: unknown;
      };
  title?: unknown;
  iconUrl?: unknown;
  id?: unknown;
  style?: unknown;
  fallback?: unknown;
  tooltip?: unknown;
  isEnabled?: unknown;
  mode?: unknown;
  requires?: unknown;
};
// ...
export type ImplementationsOfAction =
  | Action1
  | Action2
  | Action3
  | Action4
  | Action5;
// ...
export type Action1 = Extendable & {
  /**
   * Must be `Action.Execute`
   */
  type?: "Action.Execute";
  /**
   * The card author-defined verb associated with this action.
   */
  verb?: string;
  /**
   * Initial data that input fields will be combined with. These are essentially ‘hidden’ properties.
   */
  data?:
    | string
    | {
        [k: string]: unknown;
      };
  /**
   * Controls which inputs are associated with the action.
   */
  associatedInputs?:
    | ("Auto" | "None")
    | {
        [k: string]: unknown;
      };
  title?: unknown;
  iconUrl?: unknown;
  id?: unknown;
  style?: unknown;
  fallback?: unknown;
  tooltip?: unknown;
  isEnabled?: unknown;
  mode?: unknown;
  requires?: unknown;
};
/**
 * When invoked, show the given url either by launching it in an external web browser or showing within an embedded web browser.
 */
export type Action2 = Extendable & {
  /**
   * Must be `Action.OpenUrl`
   */
  type?: "Action.OpenUrl";
  /**
   * The URL to open.
   */
  url: string;
  title?: unknown;
  iconUrl?: unknown;
  id?: unknown;
  style?: unknown;
  fallback?: unknown;
  tooltip?: unknown;
  isEnabled?: unknown;
  mode?: unknown;
  requires?: unknown;
};
/**
 * Defines an AdaptiveCard which is shown to the user when the button or link is clicked.
 */
export type Action3 = Extendable & {
  /**
   * Must be `Action.ShowCard`
   */
  type?: "Action.ShowCard";
  card?: AdaptiveCard1;
  title?: unknown;
  iconUrl?: unknown;
  id?: unknown;
  style?: unknown;
  fallback?: unknown;
  tooltip?: unknown;
  isEnabled?: unknown;
  mode?: unknown;
  requires?: unknown;
};
export type ImplementationsOfElement =
  | ActionSet
  | ColumnSet
  | Container
  | FactSet
  | Image
  | ImageSet
  | Input
  | Input2
  | Input3
  | Input4
  | Input5
  | Input6
  | Media
  | RichTextBlock
  | Table
  | TextBlock;
/**
 * Displays a set of actions.
 */
export type ActionSet = Extendable2 & {
  /**
   * Must be `ActionSet`
   */
  type?: "ActionSet";
  /**
   * The array of `Action` elements to show.
   */
  actions: ImplementationsOfAction[];
  fallback?: unknown;
  height?: unknown;
  separator?: unknown;
  spacing?: unknown;
  id?: unknown;
  isVisible?: unknown;
  requires?: unknown;
};
// ...
export type ColumnSet = Extendable2 & {
  /**
   * Must be `ColumnSet`
   */
  type?: "ColumnSet";
  /**
   * The array of `Columns` to divide the region into.
   */
  columns?: Column[];
  /**
   * An Action that will be invoked when the `ColumnSet` is tapped or selected. `Action.ShowCard` is not supported.
   */
  selectAction?: Action1 | Action2 | Action4 | Action5;
  /**
   * Style hint for `ColumnSet`.
   */
  style?: ContainerStyle | null;
  /**
   * Determines whether the element should bleed through its parent's padding.
   */
  bleed?: boolean;
  /**
   * Specifies the minimum height of the column set in pixels, like `"80px"`.
   */
  minHeight?: string;
  /**
   * Controls the horizontal alignment of the ColumnSet. When not specified, the value of horizontalAlignment is inherited from the parent container. If no parent container has horizontalAlignment set, it defaults to Left.
   */
  horizontalAlignment?: HorizontalAlignment | null;
  fallback?: unknown;
  height?: unknown;
  separator?: unknown;
  spacing?: unknown;
  id?: unknown;
  isVisible?: unknown;
  requires?: unknown;
};
// ...
export type Action4 = Extendable & {
  /**
   * Must be `Action.Submit`
   */
  type?: "Action.Submit";
  /**
   * Initial data that input fields will be combined with. These are essentially ‘hidden’ properties.
   */
  data?:
    | string
    | {
        [k: string]: unknown;
      };
  /**
   * Controls which inputs are associated with the submit action.
   */
  associatedInputs?:
    | ("Auto" | "None")
    | {
        [k: string]: unknown;
      };
  title?: unknown;
  iconUrl?: unknown;
  id?: unknown;
  style?: unknown;
  fallback?: unknown;
  tooltip?: unknown;
  isEnabled?: unknown;
  mode?: unknown;
  requires?: unknown;
};
/**
 * An action that toggles the visibility of associated card elements.
 */
export type Action5 = Extendable & {
  /**
   * Must be `Action.ToggleVisibility`
   */
  type?: "Action.ToggleVisibility";
  /**
   * The array of TargetElements. It is not recommended to include Input elements with validation under Action.Toggle due to confusion that can arise from invalid inputs that are not currently visible. See https://docs.microsoft.com/en-us/adaptive-cards/authoring-cards/input-validation for more information.
   */
  targetElements: TargetElement[];
  title?: unknown;
  iconUrl?: unknown;
  id?: unknown;
  style?: unknown;
  fallback?: unknown;
  tooltip?: unknown;
  isEnabled?: unknown;
  mode?: unknown;
  requires?: unknown;
};
// ...
export type Container = Extendable2 & {
  /**
   * Must be `Container`
   */
  type?: "Container";
  /**
   * The card elements to render inside the `Container`.
   */
  items: ImplementationsOfElement[];
  /**
   * An Action that will be invoked when the `Container` is tapped or selected. `Action.ShowCard` is not supported.
   */
  selectAction?: Action1 | Action2 | Action4 | Action5;
  /**
   * Style hint for `Container`.
   */
  style?: ContainerStyle | null;
  /**
   * Defines how the content should be aligned vertically within the container. When not specified, the value of verticalContentAlignment is inherited from the parent container. If no parent container has verticalContentAlignment set, it defaults to Top.
   */
  verticalContentAlignment?: VerticalContentAlignment | null;
  /**
   * Determines whether the element should bleed through its parent's padding.
   */
  bleed?: boolean;
  /**
   * Specifies the background image. Acceptable formats are PNG, JPEG, and GIF
   */
  backgroundImage?: BackgroundImage | string;
  /**
   * Specifies the minimum height of the container in pixels, like `"80px"`.
   */
  minHeight?: string;
  /**
   * When `true` content in this container should be presented right to left. When 'false' content in this container should be presented left to right. When unset layout direction will inherit from parent container or column. If unset in all ancestors, the default platform behavior will apply.
   */
  "rtl?"?: boolean | null;
  fallback?: unknown;
  height?: unknown;
  separator?: unknown;
  spacing?: unknown;
  id?: unknown;
  isVisible?: unknown;
  requires?: unknown;
};
/**
 * The FactSet element displays a series of facts (i.e. name/value pairs) in a tabular form.
 */
export type FactSet = Extendable2 & {
  /**
   * Must be `FactSet`
   */
  type?: "FactSet";
  /**
   * The array of `Fact`'s.
   */
  facts: Fact[];
  fallback?: unknown;
  height?: unknown;
  separator?: unknown;
  spacing?: unknown;
  id?: unknown;
  isVisible?: unknown;
  requires?: unknown;
};
// ...
export interface AdaptiveCard {
  /**
   * Must be `AdaptiveCard`
   */
  type?: "AdaptiveCard";
  /**
   * Schema version that this card requires. If a client is **lower** than this version, the `fallbackText` will be rendered. NOTE: Version is not required for cards within an `Action.ShowCard`. However, it *is* required for the top-level card.
   */
  version?: string;
  refresh?: Refresh;
  authentication?: Authentication;
  /**
   * The card elements to show in the primary card region.
   */
  body?: ImplementationsOfElement[];
  /**
   * The Actions to show in the card's action bar.
   */
  actions?: ImplementationsOfAction[];
  /**
   * An Action that will be invoked when the card is tapped or selected. `Action.ShowCard` is not supported.
   */
  selectAction?: Action1 | Action2 | Action4 | Action5;
  /**
   * Text shown when the client doesn't support the version specified (may contain markdown).
   */
  fallbackText?: string;
  /**
   * Specifies the background image of the card.
   */
  backgroundImage?: BackgroundImage | string;
  /**
   * Specifies the minimum height of the card.
   */
  minHeight?: string;
  /**
   * When `true` content in this Adaptive Card should be presented right to left. When 'false' content in this Adaptive Card should be presented left to right. If unset, the default platform behavior will apply.
   */
  rtl?: boolean | null;
  /**
   * Specifies what should be spoken for this entire card. This is simple text or SSML fragment.
   */
  speak?: string;
  /**
   * The 2-letter ISO-639-1 language used in the card. Used to localize any date/time functions.
   */
  lang?: string;
  /**
   * Defines how the content should be aligned vertically within the container. Only relevant for fixed-height cards, or cards with a `minHeight` specified.
   */
  verticalContentAlignment?:
    | ("top" | "center" | "bottom")
    | {
        [k: string]: unknown;
      };
  /**
   * The Adaptive Card schema.
   */
  $schema?: string;
}
/**
 * Defines how the card can be refreshed by making a request to the target Bot.
 */
export interface Refresh {
  /**
   * Must be `Refresh`
   */
  type?: "Refresh";
  action?: Action;
  /**
   * A list of user Ids informing the client for which users should the refresh action should be run automatically. Some clients will not run the refresh action automatically unless this property is specified. Some clients may ignore this property and always run the refresh action automatically.
   */
  userIds?: string[];
}
// ...
export interface Authentication {
  /**
   * Must be `Authentication`
   */
  type?: "Authentication";
  /**
   * Text that can be displayed to the end user when prompting them to authenticate.
   */
  text?: string;
  /**
   * The identifier for registered OAuth connection setting information.
   */
  connectionName?: string;
  tokenExchangeResource?: TokenExchangeResource;
  /**
   * Buttons that should be displayed to the user when prompting for authentication. The array MUST contain one button of type "signin". Other button types are not currently supported.
   */
  buttons?: AuthCardButton[];
}
/**
 * Provides information required to enable on-behalf-of single sign-on user authentication.
 */
export interface TokenExchangeResource {
  /**
   * Must be `TokenExchangeResource`
   */
  type?: "TokenExchangeResource";
  /**
   * The unique identified of this token exchange instance.
   */
  id: string;
  /**
   * An application ID or resource identifier with which to exchange a token on behalf of. This property is identity provider- and application-specific.
   */
  uri: string;
  /**
   * An identifier for the identity provider with which to attempt a token exchange.
   */
  providerId: string;
}
/**
 * Defines a button as displayed when prompting a user to authenticate. This maps to the cardAction type defined by the Bot Framework (https://docs.microsoft.com/dotnet/api/microsoft.bot.schema.cardaction).
 */
export interface AuthCardButton {
  /**
   * The type of the button.
   */
  type: string;
  /**
   * The caption of the button.
   */
  title?: string;
  /**
   * A URL to an image to display alongside the button's caption.
   */
  image?: string;
  /**
   * The value associated with the button. The meaning of value depends on the button's type.
   */
  value: string;
}
/**
 * Specifies a background image. Acceptable formats are PNG, JPEG, and GIF
 */
export interface BackgroundImage {
  /**
   * Must be `BackgroundImage`
   */
  type?: "BackgroundImage";
  /**
   * The URL (or data url) of the image. Acceptable formats are PNG, JPEG, and GIF
   */
  url: string;
  /**
   * Describes how the image should fill the area.
   */
  fillMode?:
    | ("cover" | "repeatHorizontally" | "repeatVertically" | "repeat")
    | {
        [k: string]: unknown;
      };
  /**
   * Describes how the image should be aligned if it must be cropped or if using repeat fill mode.
   */
  horizontalAlignment?:
    | ("left" | "center" | "right")
    | {
        [k: string]: unknown;
      };
  /**
   * Describes how the image should be aligned if it must be cropped or if using repeat fill mode.
   */
  verticalAlignment?:
    | ("top" | "center" | "bottom")
    | {
        [k: string]: unknown;
      };
}